requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.2
colorama==0.4.6
curl_cffi==0.12.0
dash==3.2.0
dash-bootstrap-components==2.0.3
dash-bootstrap-templates==2.1.0
Flask==3.1.1
fred==3.1
fredapi==0.5.2
frozendict==2.4.6
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multitasking==0.0.12
narwhals==2.0.1
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
peewee==3.18.2
platformdirs==4.3.8
plotly==6.2.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
retrying==1.4.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.42
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
xlsxwriter==3.2.5
yfinance==0.2.65
zipp==3.23.0
